Add buffer option to filters.splitter.
New option filters.splitter.buffer (double, default 0).
Allow for a positive buffer around each square cell.
If buffer is greater than zero, there is the possibility
that some points within the cloud will be present in multiple
output clouds.

The default origin for filters.splitter is no longer the first
point in the cloud, but rather the minimum bounds of the entire
cloud. This will likely make the splitter unit test fail.